# backend/doc/api-docs.yml
swagger: "2.0"
info:
  title: PersonalMediaVault Backend API
  description: API to access PersonalMediaVault from a web client.
  version: 2.0.0

# cSpell:ignore allof, anyof, noneof, gyan
schemes:
  - http
  - https
securityDefinitions:
  SessionTokenHeader:
    type: apiKey
    in: header
    name: x-session-token
...
paths:
...
  "/api/media/{id}/albums":
    get:
      summary: Gets media albums list
      description: Gets media albums list
      tags:
        - media
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: number
          format: int64
          description: Media ID
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        404:
          description: Not found
        200:
          description: Albums list
          schema:
            type: array
            items:
              type: number
              format: int64
              description: Album ID
...
  "/api/albums":
    get:
      summary: Get albums
      description: Gets albums stored in the vault
      tags:
        - albums
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: "mode"
          type: string
          enum: 
            - default
            - min
          description: "Set to min to omit size and thumbnail"
          required: false
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        200:
          description: Albums
          schema:
            type: array
            items:
              properties:
                id:
                  type: number
                  format: int64
                  description: Album ID
                name:
                  type: string
                  description: "Album name"
                  example: "Album name"
                size:
                  type: number
                  format: int32
                  description: Number of items in the album
                thumbnail:
                  type: string
                  description: Thumbnail for the album (of the first media)
                lm:
                  type: number
                  format: int64
                  description: Last modified timestamp (Unix milliseconds)
                    
    post:
      summary: Create album
      description: Gets albums stored in the vault
      tags:
        - albums
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: Album name
          schema:
            properties:
              name:
                type: string
                description: Album name. Can't be blank. Max 255 chars.
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request.
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * INVALID_NAME - Invalid album name
                enum:
                  - INVALID_NAME
              message:
                type: string
                description: "Error message"
                example: "Error message"
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
        200:
          description: Album created
          schema:
            properties:
              album_id:
                type: number
                format: int64
                description: Album ID
  "/api/albums/{id}":
    get:
      summary: Get album
      description: Gets album information
      tags:
        - albums
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          type: number
          format: int64
          name: id
          description: Album ID
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        404:
          description: Not found
        200:
          description: Album information
          schema:
            properties:
              id:
                type: number
                format: int64
                description: Album ID
              name:
                type: string
                description: "Album name"
                example: "Album name"
              list:
                type: array
                description: List of media assets of the album, in order.
                items:
                  $ref: '#/definitions/MediaListItem'
              lm:
                type: number
                format: int64
                description: Last modified timestamp (Unix milliseconds)
              thumbnail:
                type: string
                description: Thumbnail for the album (of the first media)
...
definitions:
    MediaListItem:
      description: Media asset info to show in lists
      properties:
        id:
          type: number
          format: int64
          description: "Media ID"
        type:
          type: number
          format: int32
          description: >
            Media Type:
              * 0 - Deleted media
              * 1 - Image
              * 2 - Video
              * 3 - Audio
          enum:
            - 0
            - 1
            - 2
            - 3
        title:
          type: string
          description: "Media title"
          example: "Media title"
        description:
          type: string
          description: "Media description"
          example: "Media description"
        tags:
          type: array
          description: List of tags
          items:
            type: number
            format: int64
        thumbnail:
          type: string
          description: "Path to reach the thumbnail. Empty if no thumbnail. Thumbnails are squared images."
          example: "/assets/b/0/0/thumbnail.jpg"
        duration:
          type: number
          format: double
          description: Duration of the media in seconds (videos and audios)
...